backend/requirements.txt
fastapi==0.123.4 
uvicorn[standard]==0.38.0
pytesseract==0.3.13 
Pillow==12.0.0
discord.py==2.6.4
python-dotenv==1.2.1
aiohttp==3.13.2
audioop-lts==0.2.2
opencv-python==4.12.0.88
numpy==2.2.6
easyocr==1.7.2
boto3==1.42.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic==2.10.5
pydantic-settings==2.7.1